refactor: transformando páginas para nome e diretório correto
